+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * feat: instance caching for collection decorators
 * fix: should be static instead of self for instantiation in decorators
 * feat: instance caching for criteria recollections
+* test: criteria recollection
 
 ## 0.3.0